exif: fix a c++ compiler warning.

+++ src/common/exif.cc
@@ -1781,7 +1781,7 @@ int dt_exif_thumbnail(
     if(!dt_imageio_jpeg_decompress_header(buf.pData_, buf.size_, &jpg))
     {
       // don't upsample those:
-      if(jpg.width < width || jpg.height < height) return 1;
+      if((uint32_t)jpg.width < width || (uint32_t)jpg.height < height) return 1;
       if(!y_beg && !y_end)
       { // if those weren't set, do it now:
         y_beg = 0;
